Updated floating point standards.

+++ lioamber/fileio/movieprint.f90
@@ -1,14 +1,15 @@
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%!
+#include "../complex_type.fh"
 subroutine movieprint( natoms, mbasis, nstep, nucids, nucpos, elecdens )
 
    use fileio_data, only: movie_count, movie_nfreq, movie_name0
    implicit none
-   integer   , intent(in) :: natoms
-   integer   , intent(in) :: mbasis
-   integer   , intent(in) :: nstep
-   integer   , intent(in) :: nucids(natoms)
-   real*8    , intent(in) :: nucpos(natoms,3)
-   complex*16, intent(in) :: elecdens(mbasis,mbasis)
+   integer        , intent(in) :: natoms
+   integer        , intent(in) :: mbasis
+   integer        , intent(in) :: nstep
+   integer        , intent(in) :: nucids(natoms)
+   real(kind=8)   , intent(in) :: nucpos(natoms,3)
+   complex(kind=8), intent(in) :: elecdens(mbasis,mbasis)
 
    character(len=3), allocatable :: nucids_name(:)
    character(len=4)  :: charnum
